+++ docs/pages/locality/index.html
@@ -8,7 +8,11 @@
     <script src="../../mainmapdata.js"></script>
     <script src="../../statemap.js"></script>
     <script type="module" src="bundle.js"></script>
-    <link rel="stylesheet" type="text/css" href="../../style.css">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous" />
+    <script defer src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+        crossorigin="anonymous"></script>
     <script>
         (function (i, s, o, g, r, a, m) {
             i['GoogleAnalyticsObject'] = r; i[r] = i[r] || function () {
@@ -24,15 +28,37 @@
 
 </head>
 
-<body>
-
-    <h2><span id="localityText"></span></h2>
-    <br><br>
-    <a href="../../index.html">Return to All Providers Map</a>.
-    <br><br>
-    <div id="coveragemap"></div>
-    <h2>Providers Serving This Locality:</h2><br>
-    <ul id="providerList" style='list-style: none;'>
+<body class="p-3 m-0 border-0 bd-example m-0 border-0">
+    <nav class="navbar navbar-expand-lg bg-body-tertiary">
+        <div class="container-fluid">
+            <a class="navbar-brand" href="../../index.html">VA DJJ Provider Directory</a>
+            <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
+                data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false"
+                aria-label="Toggle navigation">
+                <span class="navbar-toggler-icon"></span>
+            </button>
+            <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+                    <li class="nav-item">
+                        <a class="nav-link" aria-current="page" href="#Providers">Providers</a>
+                    </li>
+                </ul>
+            </div>
+        </div>
+    </nav>
+    <h1 class="dispaly-1"><span id="localityText"></span></h1>
+    <div class="container mt-5">
+        <div class="row mx-auto">
+            <div class="col">
+                <div id="coveragemap"></div>
+            </div>
+        </div>
+    </div>
+    <section class="container mt-5">
+        <a id="Providers"></a>
+        <h2>Providers Serving This Locality:</h2>
+        <div id="providerList" class="row row-cols-3"></div>
+    </section>
 </body>
 
 </html>
