+++ src/pages/csu/index.js
@@ -5,36 +5,12 @@ import {
   providerServiceList,
   serviceProvidersList,
 } from "./api.js";
+import { headingList } from "../../lib/utils.js";
 
 const csuNameSpan = document.getElementById("csuName");
 if (csuNameSpan) {
   csuNameSpan.innerHTML = thisCSU?.name || "CSU";
 }
-
-const headingList = (headingContent, listItems) => {
-  const providerCard = document.createElement("div");
-  providerCard.className = "card";
-  const cardHeader = document.createElement("div");
-  cardHeader.className = "card-header";
-
-  const providerTitle = document.createElement("h5");
-  providerTitle.className = "card-title";
-  providerTitle.innerHTML = headingContent;
-  cardHeader.appendChild(providerTitle);
-
-  const list = document.createElement("ul");
-  list.className = "list-group-flush mt-2";
-  listItems.forEach((item) => {
-    const li = document.createElement("li");
-    li.className = "list-group-item";
-    li.innerText = item;
-    list.appendChild(li);
-  });
-
-  providerCard.appendChild(cardHeader);
-  providerCard.appendChild(list);
-  return providerCard;
-};
 
 const providerDiv = document.getElementById("providerList");
 CSUProviders.forEach((provider) => {
